+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
 - Remove `node` user (Always run with host user id)
 - Merge `RUN`'s in `Dockerfile`
 - Disable npm audit/fund
+- Update `flux-build-utils`
 
 ## v2024-01-15-1